refactor: cheatsheet and whatsnew module
